refactor: set up new pkg for trafficproxyclient
